+++ hutool-core/src/main/java/cn/hutool/core/comparator/FieldsComparator.java
@@ -1,7 +1,7 @@
 package cn.hutool.core.comparator;
 
+import cn.hutool.core.lang.Assert;
 import cn.hutool.core.util.ClassUtil;
-import cn.hutool.core.util.StrUtil;
 
 import java.lang.reflect.Field;
 
@@ -12,10 +12,8 @@ import java.lang.reflect.Field;
  * @param <T> 被比较的Bean
  * @author Looly
  */
-public class FieldsComparator<T> extends BaseFieldComparator<T> {
+public class FieldsComparator<T> extends NullComparator<T> {
 	private static final long serialVersionUID = 8649196282886500803L;
-
-	private final Field[] fields;
 
 	/**
 	 * 构造
@@ -24,23 +22,29 @@ public class FieldsComparator<T> extends BaseFieldComparator<T> {
 	 * @param fieldNames 多个字段名
 	 */
 	public FieldsComparator(Class<T> beanClass, String... fieldNames) {
-		this.fields = new Field[fieldNames.length];
-		for (int i = 0; i < fieldNames.length; i++) {
-			this.fields[i] = ClassUtil.getDeclaredField(beanClass, fieldNames[i]);
-			if (this.fields[i] == null) {
-				throw new IllegalArgumentException(StrUtil.format("Field [{}] not found in Class [{}]", fieldNames[i], beanClass.getName()));
-			}
-		}
+		this(true, beanClass, fieldNames);
 	}
 
-	@Override
-	public int compare(T o1, T o2) {
-		for (Field field : fields) {
-			int compare = this.compareItem(o1, o2, field);
-			if (compare != 0) {
-				return compare;
+	/**
+	 * 构造
+	 *
+	 * @param nullGreater 是否{@code null}在后
+	 * @param beanClass   Bean类
+	 * @param fieldNames  多个字段名
+	 */
+	public FieldsComparator(boolean nullGreater, Class<T> beanClass, String... fieldNames) {
+		super(nullGreater, (a, b) -> {
+			Field field;
+			for (String fieldName : fieldNames) {
+				field = ClassUtil.getDeclaredField(beanClass, fieldName);
+				Assert.notNull(field, "Field [{}] not found in Class [{}]", fieldName, beanClass.getName());
+				final int compare = new FieldComparator<>(field).compare(a, b);
+				if (0 != compare) {
+					return compare;
+				}
 			}
-		}
-		return 0;
+			return 0;
+		});
 	}
+
 }